requirements.txt
catbox_api==1.3
py-cord==2.6.4
python-dotenv==1.2.1
Requests==2.32.5
